Even cleaner.
