+++ src/html/listener/stdweb.rs
@@ -1,52 +1,3 @@
-use super::*;
-use crate::callback::Callback;
-use crate::virtual_dom::Listener;
-#[allow(unused_imports)]
-use stdweb::web::EventListenerHandle;
-
-macro_rules! impl_action {
-    ($($action:ident($event:ident : $type:ident) -> $ret:ty => $convert:expr)*) => {$(
-        /// An abstract implementation of a listener.
-        pub mod $action {
-            use stdweb::web::{IEventTarget, Element};
-            use stdweb::web::event::{IEvent, $type};
-            use super::*;
-
-            /// A wrapper for a callback which attaches event listeners to elements.
-            #[derive(Clone, Debug)]
-            pub struct Wrapper {
-                callback: Callback<Event>,
-            }
-
-            impl Wrapper {
-                /// Create a wrapper for an event-typed callback
-                pub fn new(callback: Callback<Event>) -> Self {
-                    Wrapper { callback }
-                }
-            }
-
-            /// And event type which keeps the returned type.
-            pub type Event = $ret;
-
-            impl Listener for Wrapper {
-                fn kind(&self) -> &'static str {
-                    stringify!($action)
-                }
-
-                fn attach(&self, element: &Element) -> EventListenerHandle {
-                    let this = element.clone();
-                    let callback = self.callback.clone();
-                    let listener = move |event: $type| {
-                        event.stop_propagation();
-                        callback.emit($convert(&this, event));
-                    };
-                    element.add_event_listener(listener)
-                }
-            }
-        }
-    )*};
-}
-
 // Inspired by: http://package.elm-lang.org/packages/elm-lang/html/2.0.0/Html-Events
 impl_action! {
     onclick(event: ClickEvent) -> ClickEvent => |_, event| { event }
