+++ src/html/listener/web_sys.rs
@@ -1,67 +1,3 @@
-use super::*;
-use crate::callback::Callback;
-use crate::compat::EventListenerHandle;
-use crate::virtual_dom::Listener;
-
-macro_rules! impl_action {
-    ($($action:ident(name: $name:literal, event: $type:ident) -> $ret:ty => $convert:expr)*) => {$(
-        /// An abstract implementation of a listener.
-        pub mod $action {
-            use std::mem::ManuallyDrop;
-            use wasm_bindgen::{closure::Closure, JsCast};
-            use web_sys::{Element, EventTarget, $type as WebSysType};
-            use super::*;
-
-            /// A wrapper for a callback which attaches event listeners to elements.
-            #[derive(Clone, Debug)]
-            pub struct Wrapper {
-                callback: Callback<Event>,
-            }
-
-            impl Wrapper {
-                /// Create a wrapper for an event-typed callback
-                pub fn new(callback: Callback<Event>) -> Self {
-                    Wrapper { callback }
-                }
-            }
-
-            /// And event type which keeps the returned type.
-            pub type Event = $ret;
-
-            impl Listener for Wrapper {
-                fn kind(&self) -> &'static str {
-                    stringify!($action)
-                }
-
-                fn attach(&self, element: &Element) -> EventListenerHandle {
-                    let this = element.clone();
-                    let callback = self.callback.clone();
-                    let listener = move |event: web_sys::Event| {
-                        event.stop_propagation();
-                        let event = event.dyn_into::<WebSysType>().expect("wrong event type");
-                        callback.emit($convert(&this, event));
-                    };
-
-                    let target = EventTarget::from(element.clone());
-                    let listener = Closure::wrap(Box::new(listener) as Box<dyn Fn(web_sys::Event)>);
-                    target
-                        .add_event_listener_with_callback(
-                            stringify!($name),
-                            listener.as_ref().unchecked_ref(),
-                        )
-                        .expect("failed to add event listener");
-
-                    return EventListenerHandle {
-                        target,
-                        r#type: stringify!($name),
-                        callback: ManuallyDrop::new(listener),
-                    }
-                }
-            }
-        }
-    )*};
-}
-
 // Inspired by: http://package.elm-lang.org/packages/elm-lang/html/2.0.0/Html-Events
 impl_action! {
     onclick(name: "click", event: MouseEvent) -> web_sys::MouseEvent => |_, event| { event }
